--- test/expect_text.txt
@@ -1,67 +1,0 @@
-The 33 Strategies of War 
-1
-The 33 Strategies of War
-The 33 Strategies of War
-Author
-Robert Greene (author)
-Country
-United States
-Language
-English
-Genre(s)
-Business, Management, Military History, Psychology, self-improvement
-Publisher
-Penguin Group (HC); HighBridge Audio (CD)
-Publication date
-January and April 2006
-Media type
-Print (Hardcover) and CD
-Pages
-496
-ISBN
-ISBN 0-670-03457-6 (HC); 978-1-59887-091-6 (CD)
-The 33 Strategies of War
-by Robert Greene is a "guide to the subtle social game of everyday life informed by the ...
-military principles in war.".
-[1]
-It is composed of discussions and examples on offensive and defensive strategies
-from a wide variety of people and conditions such as Napoleon Bonaparte, Lawrence of Arabia, Alexander the Great,
-and the Tet Offensive.
-[2]
-The scope of the book is broad, applying not only to violent conflicts but also social
-conflicts such as family quarrels and business negotiations.
-The book is di
-vided into
-five parts: Self-Directed Warfare, Organizational (Team) Warfare, Defensive Warfare,
-Offensive Warfare and Unconventional (Dirty) Warfare.
-[1]
-Each part contains a differing number of strategies, each
-in a chapter. Each chapter has a similar layout. Descriptions of battles, political and business situations are
-accompanied by Greene's interpretation. There are occasional instructional sections followed by examples. All
-chapters end with a "Reversal" to give a brief discussion of where the strategy may not apply, a contrary view or
-defense. Throughout the book Mr. Greene includes quotes from a variety of sources. These are incorporated in the
-margins and between sections.
-Although one reviewer has called the book "an indispensable book, [which] provides all the psychological
-ammunition you need to overcome patterns of failure and forever gain the upper hand,"
-[3]
-another one found it
-"perplexing
- —
-if not downright unhealthy
- —
-[to publish] a book on the lessons of war for everything but war at a
-time when we [Canada] are, er, at war."
-[4]
-Yet another reviewer found the book's coverage of military history
-informative, but the political tales "mostly foolish or just plain wrong".
-[5]
-On Amazon.com, the majority of the
-reviews on the book are positive.
-The 33 Strategies of War 
-was part of the reading list for youths attending the Indigenous Leadership Forum
-organised by the University of Victoria, which aimed to redesign radical Indigenous politics and the Indigenist
-movement.
-[6]
-It is also read by students attending a Southwestern Baptist Theological Seminary course in Christian
-apologetics.
-[7]
